+++ src/extra/calibration.py
@@ -4,7 +4,7 @@ import json
 import re
 from collections.abc import Mapping
 from dataclasses import dataclass, field, fields, replace
-from datetime import date, datetime, time, timezone
+from datetime import date, datetime, time, timedelta, timezone
 from enum import IntFlag
 from fnmatch import fnmatch
 from functools import lru_cache
@@ -16,6 +16,7 @@ from urllib.parse import urljoin
 from warnings import warn
 
 import h5py
+import numpy as np
 import pasha as psh
 import requests
 from extra_data.read_machinery import find_proposal
@@ -314,7 +315,7 @@ def _summarise_mod_names(l):
     E.g. ['LPD00', 'LPD01', 'LPD03'] -> ['LPD00', 'LPD01'], ['LPD03']
     """
     grp = []
-    num_re = re.compile('(\d+)')
+    num_re = re.compile(r'(\d+)')
 
     def extends_grp(split):
         if len(split) != len(grp[0]):
@@ -1347,6 +1348,68 @@ class LPDConditions(ConditionsBase):
             del cond["Parallel gain"]
 
         return cond
+
+
+def lpd_dark_consts_with_fallback(
+    condition: "LPDConditions",
+    detector_name,
+    event_at=None,
+    preference_time=timedelta(days=7),
+    **kwargs
+):
+    """Look up LPD dark constants with fallback to constants for all memory cells
+
+    The parameters are mostly the same as for CalibrationData.from_condition().
+    Constants with the matching memory cell order will be used if they're closer
+    in time than the fallback, or up to *preference_time* (default 5 days)
+    further.
+    """
+    if event_at is None:
+        event_at = datetime.now(timezone.utc)
+    elif isinstance(event_at, str):
+        event_at = datetime.fromisoformat(event_at)
+
+    cd_preferred = CalibrationData.from_condition(
+        condition,
+        detector_name,
+        calibrations=["Offset", "Noise", "BadPixelsDark"],
+        **kwargs
+    )
+
+    fallback_mem_cell_order = ",".join([str(i) for i in range(510)]) + ","
+    if condition.memory_cell_order == fallback_mem_cell_order:
+        return cd_preferred  # Fallback would be the same
+
+    cd_fallback = CalibrationData.from_condition(
+        replace(condition, memory_cell_order=fallback_mem_cell_order),
+        detector_name,
+        calibrations=["Offset", "Noise", "BadPixelsDark"],
+        **kwargs
+    )
+
+    def time_distance(cd, aggregator):
+        bva = cd['Offset', aggregator].metadata("begin_validity_at")
+        return abs(datetime.fromisoformat(bva) - event_at)
+
+    # Select which modules to use from the fallback. In most cases, this will
+    # be either none or all, as we'll have the same timestamps for all modules,
+    # but it may vary if a module is replaced. In theory the timestamps can also
+    # be different for each calibration type, but that's unlikely, so we just
+    # look at the offset timestamps.
+    aggrs_fallback = []
+    for da in (set(cd_preferred.aggregator_names) | set(cd_fallback.aggregator_names)):
+        if da not in cd_fallback.get('Offset', {}):
+            continue
+        elif da not in cd_preferred.get('Offset', {}):
+            aggrs_fallback.append(da)  # Only in fallback
+        else:
+            closer_by = time_distance(cd_preferred, da) - time_distance(cd_fallback, da)
+            if closer_by > preference_time:
+                aggrs_fallback.append(da)
+
+    return cd_preferred.merge(
+        cd_fallback.select_modules(aggregator_names=aggrs_fallback)
+    )
 
 
 @dataclass